Moved SocialLoginButtons from global components to /user/_components
